refactor clear / select
