+++ odpc.client/src/features/publicatie/PublicatieDetails.vue
@@ -145,7 +145,7 @@
 </template>
 
 <script setup lang="ts">
-import { computed, ref, watch } from "vue";
+import { computed, ref } from "vue";
 import { useRouter } from "vue-router";
 import SimpleSpinner from "@/components/SimpleSpinner.vue";
 import AlertInline from "@/components/AlertInline.vue";
@@ -224,64 +224,6 @@ const {
 // Permissions
 const { isReadonly, canDraft, canDelete, canRetract, unauthorized, groepWaardelijsten } =
   usePublicatiePermissions(publicatie, mijnGebruikersgroepen);
-
-// Externally created publicaties will not have a eigenaarGroep untill updated from ODPC
-const isPublicatieWithoutEigenaarGroep = computed(
-  () => !!publicatie.value.uuid && !publicatie.value.eigenaarGroep
-);
-
-const presetSingleGebruikersgroep = () => {
-  if (mijnGebruikersgroepen.value?.length !== 1) return;
-
-  const { uuid, naam } = mijnGebruikersgroepen.value[0];
-
-  publicatie.value.eigenaarGroep = { identifier: uuid, weergaveNaam: naam };
-};
-
-const presetSingleWaardelijstOption = () => {
-  const { organisaties, informatiecategorieen } = groepWaardelijsten.value;
-
-  if (organisaties?.length === 1) publicatie.value.publisher = organisaties[0].uuid;
-
-  if (informatiecategorieen?.length === 1)
-    publicatie.value.informatieCategorieen.push(informatiecategorieen[0].uuid);
-};
-
-// Check to do some presets after all loading is finished
-watch(isLoading, () => {
-  if (hasError.value) return;
-
-  // Preset eigenaarGroep of a new - or externally created publicatie when only one mijnGebruikersgroep
-  if (!publicatie.value.uuid || isPublicatieWithoutEigenaarGroep.value) {
-    presetSingleGebruikersgroep();
-  }
-
-  // Preset publisher and informatiecategorie of a new publicatie when only one option available
-  if (!publicatie.value.uuid) {
-    presetSingleWaardelijstOption();
-  }
-});
-
-// Clear waardelijsten of publicatie when mismatch waardelijsten gebruikersgroep (unauthorized) on
-// a) change from one to another eigenaarGroep or
-// b) initial select eigenaarGroep when isPublicatieWithoutEigenaarGroep
-watch(
-  () => publicatie.value.eigenaarGroep,
-  (_, oldGroep) => {
-    if (!unauthorized.value) return;
-
-    if (!!oldGroep || isPublicatieWithoutEigenaarGroep.value) {
-      publicatie.value = {
-        ...publicatie.value,
-        ...{
-          publisher: "",
-          informatieCategorieen: [],
-          onderwerpen: []
-        }
-      };
-    }
-  }
-);
 
 const navigate = () => {
   if (
